NICE: Feedback ID 9737956, SRD | 19C5026i | Entitlement | research.com.apple.license-to-operate
Thanks for your feedback, and for your patience.  Apologies for the delay. We believe this has been addressed.
We made launchd debuggable via a kernel patch, which should resolve this issue, shipped with kc tutorial on researcher GitHub.
Please close if resolved, or let us know if still an issue for you.
Thank you.